+++ com.unity.perception/Runtime/GroundTruth/SimulationState.cs
@@ -67,10 +67,10 @@ namespace UnityEngine.Perception.GroundTruth
         }
 
         //A sensor will be triggered if sequenceTime is within includeThreshold seconds of the next trigger
-        const float k_IncludeInFrameThreshold = .01f;
+        const float k_SimulationTimingAccuracy = 0.01f;
         const int k_MinPendingCapturesBeforeWrite = 150;
         const int k_MinPendingMetricsBeforeWrite = 150;
-        const int k_MaxDeltaTime = 10;
+        const float k_MaxDeltaTime = 100f;
 
         public SimulationState(string outputDirectory)
         {
@@ -145,10 +145,14 @@ namespace UnityEngine.Perception.GroundTruth
         {
             public string modality;
             public string description;
-            public float period;
             public float firstCaptureTime;
+            public CaptureTriggerMode captureTriggerMode;
+            public float renderingDeltaTime;
+            public int framesBetweenCaptures;
+            public bool manualSensorAffectSimulationTiming;
 
-            public float sequenceTimeNextCapture;
+            public float sequenceTimeOfNextCapture;
+            public float sequenceTimeOfNextRender;
             public int lastCaptureFrameCount;
             public EgoHandle egoHandle;
         }
@@ -354,7 +358,8 @@ namespace UnityEngine.Perception.GroundTruth
             foreach (var kvp in m_Sensors.ToArray())
             {
                 var sensorData = kvp.Value;
-                sensorData.sequenceTimeNextCapture = SequenceTimeOfNextCapture(sensorData);
+                sensorData.sequenceTimeOfNextCapture = GetSequenceTimeOfNextCapture(sensorData);
+                sensorData.sequenceTimeOfNextRender = 0;
                 m_Sensors[kvp.Key] = sensorData;
             }
 
@@ -369,29 +374,36 @@ namespace UnityEngine.Perception.GroundTruth
             m_LastTimeScale = Time.timeScale;
         }
 
-        public void AddSensor(EgoHandle egoHandle, string modality, string description, float period, float firstCaptureTime, SensorHandle sensor)
+        public void AddSensor(EgoHandle egoHandle, string modality, string description, float firstCaptureFrame, CaptureTriggerMode captureTriggerMode, float renderingDeltaTime, int framesBetweenCaptures, bool manualSensorAffectSimulationTiming, SensorHandle sensor)
         {
             var sensorData = new SensorData()
             {
                 modality = modality,
                 description = description,
-                period = period,
-                firstCaptureTime = firstCaptureTime,
+                firstCaptureTime = firstCaptureFrame * renderingDeltaTime,
+                captureTriggerMode = captureTriggerMode,
+                renderingDeltaTime = renderingDeltaTime,
+                framesBetweenCaptures = framesBetweenCaptures,
+                manualSensorAffectSimulationTiming = manualSensorAffectSimulationTiming,
                 egoHandle = egoHandle,
                 lastCaptureFrameCount = -1
             };
-            sensorData.sequenceTimeNextCapture = SequenceTimeOfNextCapture(sensorData);
+            sensorData.sequenceTimeOfNextCapture = GetSequenceTimeOfNextCapture(sensorData);
+            sensorData.sequenceTimeOfNextRender = 0;
             m_ActiveSensors.Add(sensor);
             m_Sensors.Add(sensor, sensorData);
             m_Ids.Add(sensor.Id);
         }
 
-        float SequenceTimeOfNextCapture(SensorData sensorData)
+        float GetSequenceTimeOfNextCapture(SensorData sensorData)
         {
             // If the first capture hasn't happened yet, sequenceTimeNextCapture field won't be valid
             if (sensorData.firstCaptureTime >= UnscaledSequenceTime)
-                return sensorData.firstCaptureTime;
-            return sensorData.sequenceTimeNextCapture;
+            {
+                return sensorData.captureTriggerMode == CaptureTriggerMode.Scheduled? sensorData.firstCaptureTime : float.MaxValue;
+            }
+
+            return sensorData.sequenceTimeOfNextCapture;
         }
 
         public bool Contains(Guid id) => m_Ids.Contains(id);
@@ -435,20 +447,35 @@ namespace UnityEngine.Perception.GroundTruth
                 m_HasStarted = true;
             }
 
+            EnsureSequenceTimingsUpdated();
+
             //update the active sensors sequenceTimeNextCapture and lastCaptureFrameCount
             foreach (var activeSensor in m_ActiveSensors)
             {
                 var sensorData = m_Sensors[activeSensor];
-                if (!activeSensor.ShouldCaptureThisFrame)
-                    continue;
 
-                // TODO: AISV-845 This is an errant modification of this record that can lead to undefined behavior
-                // Leaving as-is for now because too many components depend on this logic
-                sensorData.sequenceTimeNextCapture += sensorData.period;
-                Debug.Assert(sensorData.sequenceTimeNextCapture > UnscaledSequenceTime,
-                    $"Next scheduled capture should be after {UnscaledSequenceTime} but is {sensorData.sequenceTimeNextCapture}");
-                // sensorData.sequenceTimeNextCapture = SequenceTimeOfNextCapture(sensorData);
-                sensorData.lastCaptureFrameCount = Time.frameCount;
+                if (Mathf.Abs(sensorData.sequenceTimeOfNextRender - UnscaledSequenceTime) < k_SimulationTimingAccuracy)
+                {
+                    //means this frame fulfills this sensor's simulation time requirements, we can move target to next frame.
+                    sensorData.sequenceTimeOfNextRender += sensorData.renderingDeltaTime;
+                }
+
+                if (activeSensor.ShouldCaptureThisFrame)
+                {
+                    if (sensorData.captureTriggerMode.Equals(CaptureTriggerMode.Scheduled))
+                    {
+                        sensorData.sequenceTimeOfNextCapture += sensorData.renderingDeltaTime * (sensorData.framesBetweenCaptures + 1);
+                        Debug.Assert(sensorData.sequenceTimeOfNextCapture > UnscaledSequenceTime,
+                            $"Next scheduled capture should be after {UnscaledSequenceTime} but is {sensorData.sequenceTimeOfNextCapture}");
+                    }
+                    else if (sensorData.captureTriggerMode.Equals(CaptureTriggerMode.Manual))
+                    {
+                        sensorData.sequenceTimeOfNextCapture = float.MaxValue;
+                    }
+
+                    sensorData.lastCaptureFrameCount = Time.frameCount;
+                }
+
                 m_Sensors[activeSensor] = sensorData;
             }
 
@@ -456,18 +483,44 @@ namespace UnityEngine.Perception.GroundTruth
             float nextFrameDt = k_MaxDeltaTime;
             foreach (var activeSensor in m_ActiveSensors)
             {
-                var sensorData = m_Sensors[activeSensor];
-                var thisSensorNextFrameDt = sensorData.sequenceTimeNextCapture - UnscaledSequenceTime;
+                float thisSensorNextFrameDt = -1;
 
-                Debug.Assert(thisSensorNextFrameDt > 0f, "Sensor was scheduled to run in the past but got skipped over.");
+                var sensorData = m_Sensors[activeSensor];
+                if (sensorData.captureTriggerMode.Equals(CaptureTriggerMode.Scheduled))
+                {
+                    thisSensorNextFrameDt = sensorData.sequenceTimeOfNextRender - UnscaledSequenceTime;
+
+                    Debug.Assert(thisSensorNextFrameDt > 0f, "Sensor was scheduled to capture in the past but got skipped over.");
+                }
+                else if (sensorData.captureTriggerMode.Equals(CaptureTriggerMode.Manual) && sensorData.manualSensorAffectSimulationTiming)
+                {
+                    thisSensorNextFrameDt = sensorData.sequenceTimeOfNextRender - UnscaledSequenceTime;
+                }
+
                 if (thisSensorNextFrameDt > 0f && thisSensorNextFrameDt < nextFrameDt)
                     nextFrameDt = thisSensorNextFrameDt;
+            }
+
+            if (Math.Abs(nextFrameDt - k_MaxDeltaTime) < 0.0001)
+            {
+                //means no sensor is controlling simulation timing, so we set Time.captureDeltaTime to 0 (default) which means the setting does not do anything
+                nextFrameDt = 0;
             }
 
             WritePendingCaptures();
             WritePendingMetrics();
 
             Time.captureDeltaTime = nextFrameDt;
+        }
+
+        public void SetNextCaptureTimeToNowForSensor(SensorHandle sensorHandle)
+        {
+            if (!m_Sensors.ContainsKey(sensorHandle))
+                return;
+
+            var data = m_Sensors[sensorHandle];
+            data.sequenceTimeOfNextCapture = UnscaledSequenceTime;
+            m_Sensors[sensorHandle] = data;
         }
 
         public bool ShouldCaptureThisFrame(SensorHandle sensorHandle)
@@ -479,7 +532,7 @@ namespace UnityEngine.Perception.GroundTruth
             if (data.lastCaptureFrameCount == Time.frameCount)
                 return true;
 
-            return (data.sequenceTimeNextCapture - UnscaledSequenceTime) < k_IncludeInFrameThreshold;
+            return data.sequenceTimeOfNextCapture - UnscaledSequenceTime < k_SimulationTimingAccuracy;
         }
 
         public void End()